clean up error boundary file
